+++ RELEASE_NOTES.md
@@ -17,6 +17,7 @@ Breaking changes can still occur in minor versions and patches, though, if the a
 
 * `CaseAdjustable` is a new protocol that
 * `EmojiKeyboardStyle` now lets you provide a device type in the standard style builder. 
+* `InputCallout` now lets you provide a device type in the initializer.
 * `InputSet` properties are now mutable.
 * `InputSetItem` properties are now mutable.
 * `KeyboardColorReader` is a new protocol that is implemented by `Color` and lets its implementations access keyboard colors.